pyproject.toml
[project]
name = "payments"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "sqlalchemy[asyncio]==2.0.40",
    "asyncpg==0.30.0",
    "alembic==1.15.2",
    "pydantic-settings==2.9.1",
    "faststream[rabbit]==0.5.41",
    "httpx==0.28.1",
    "aio-pika==9.5.5",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "factory-boy==3.3.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
pythonpath = ["."]
testpaths = ["tests"]
